v2/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
mysql-connector-python
redis
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-genai
fastf1
pandas
python-multipart
